Refactor access token code and tests into its own module
